refactor: split the new renderer out as AnimatedAgentRuns
This branch rewrote AnimatedDotGrid in place, which put a 1,824-line
file in the path of every other branch touching it. The new hero
renderer moves to its own component and AnimatedDotGrid returns to its
main version untouched, so both can live side by side.
